Add a configurable cache for the third-party archives
The cache system creates a global cache under
`${GEM_SUPERBUILD_CACHE_DIR}` which defaults to
`${CMAKE_BINARY_DIR/cache}`.

The existing archives from `${CMAKE_SOURCE_DIR}/binary` are copied to
the cache as they are already downloaded and can be hard to download.

All the packages can be downloaded from Internet and their integrity is
check with a SHA256 sum.

Fixes #1