Add occi-abicompat, a compatibility layer for OCCI on some platforms
occi-abicompat is a small library that converts calls to OCCI that use the
C++11 ABI of libstdc++ to calls using the C++98 ABI. This is needed if libocci
was compiled against the C++98 ABI but one wants to use the C++11 version.

Only symbols used by xDAQ::tstore are redefined.

Since the library is very small, it is added directly to the central
repository.